@@ -1,4 +1,4 @@
-import { FormGroup, FormControl } from '@angular/forms';
+import { FormGroup, FormControl, Validators } from '@angular/forms';
 
 /**
  * Registration form.
@@ -6,29 +6,50 @@ import { FormGroup, FormControl } from '@angular/forms';
 export class RegistrationForm {
     public registrationForm = new FormGroup({
         /**
-         * FIrst name input field.
+         * First Name input field.
          */
-        firstName: new FormControl(''),
+        firstName: new FormControl('', [
+            Validators.required,
+            Validators.minLength(3),
+            Validators.maxLength(255),
+        ]),
 
         /**
-         * Last name input field.
+         * Last Name input field.
          */
-        lastName: new FormControl(''),
+        lastName: new FormControl('', [
+            Validators.required,
+            Validators.minLength(3),
+            Validators.maxLength(255),
+        ]),
 
         /**
          * Email input field.
          */
-        email: new FormControl(''),
+        email: new FormControl('', [
+            Validators.email,
+            Validators.required,
+            Validators.minLength(3),
+            Validators.maxLength(255),
+        ]),
 
         /**
          * Password input field.
          */
-        password: new FormControl(''),
+        password: new FormControl('', [
+            Validators.required,
+            Validators.minLength(3),
+            Validators.maxLength(255),
+        ]),
 
         /**
-         * Confirm password input field.
+         * Confirm Password input field.
          */
-        confirmPassword: new FormControl(''),
+        confirmPassword: new FormControl('', [
+            Validators.required,
+            Validators.minLength(3),
+            Validators.maxLength(255),
+        ]),
 
         /**
          * Phone number input field.
