@@ -22,18 +22,18 @@ export class GlobalService {
   public _isLoadedData = false;
 
   /**
-   * @param _currentUser User
+   * @param _currentUser User | undefined
    */
-  public _currentUser: User;
+  public _currentUser: User | undefined;
   /**
-   * @param _avatarUrl string
+   * @param _avatarUrl string | undefined
    */
-  public _avatarUrl: string;
+  public _avatarUrl: string | undefined;
 
   /**
    * @param _roles string[]
    */
-  _roles: string[];
+  _roles: string[] = [];
   // public _roles: any[];
 
   constructor() { }
@@ -55,7 +55,10 @@ export class GlobalService {
    * @param activatedRoute ActivatedRoute
    * @returns string|null
    */
-  private _getId(idName: string, routeObject: ActivatedRouteSnapshot): string|null {
+  private _getId(
+    idName: string,
+    routeObject: ActivatedRouteSnapshot | null
+  ): string|null {
     if (!routeObject) {
         return null;
     }
