@@ -8,6 +8,7 @@ import { GalleryComponent } from '../default-pages/gallery/gallery.component';
 import { AboutComponent } from '../default-pages/about/about.component';
 import { RegistrationComponent } from '../user/registration/registration.component';
 import { AuthorizationComponent } from '../user/authorization/authorization.component';
+import { DishesModule } from '../dishes/dishes.module';
 
 const routes: Routes = [
   {
@@ -32,7 +33,7 @@ const routes: Routes = [
       },
       {
         path: 'dishes',
-        loadChildren: '../dishes/dishes.module#DishesModule'
+        loadChildren: () => DishesModule
       },
       {
         path: 'registration',
